+++ test/Resources/TestScripts/WorkflowAppFunctionApp/HttpTriggerFunction/index.js
@@ -1,0 +1,4 @@
+module.exports = function (context, req) {
+    context.res = { status: 200, body: 'ok' };
+    context.done();
+};