+++ release_notes.md
@@ -4,4 +4,5 @@
 - My change description (#PR)
 -->
 
+- Restore Workflows-bundle worker discovery on Logic App (#11759)
 - Ensure wwwroot directory exists on new slot and app creation w/ networking restrictions (#11757)